folder restructure to meet standard
