+++ src/components/footer/index.js
