+++ src/pages/index.js
@@ -5,8 +5,9 @@ import {
 } from 'react-router-dom';
 
 import Styled from 'styled-components';
-import Home from './pages/home';
-import Search from './pages/search';
+
+import Home from './home';
+import Search from './search';
 
 const StyledMain = Styled.div`
     min-height: calc(100vh - 200px);
